fix(frontend): show page-specific skeletons at route level to remove double-skeleton flash
Route-level loading.tsx files still rendered the old generic Loading
skeleton, which Next.js shows on every navigation before the page
mounts and renders its own content-shaped skeleton (added in #1852).
Users saw two different skeletons in sequence on every page change.

Extract each page's skeleton into a co-located page-skeleton.tsx and
add a per-route loading.tsx rendering the same skeleton, so navigation
shows one continuous skeleton from the first frame until content.
The group-level (protected)/loading.tsx stays as fallback for routes
without a page-specific skeleton, where it already matches the
after-mount RoleGuard fallback.

Covers dashboard, staff, staff/[id], activities, rooms, messages,
ogs-groups, students/[id], students/search, timetables, suggestions,
and the database section (loading.tsx reusing MasterDetailSkeleton).
Skeleton JSX moved verbatim; no visual changes to the skeletons
themselves.
