+++ frontend/src/app/[tenant]/(protected)/ogs-groups/page-skeleton.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { StudentCardPageSkeleton } from "~/components/students/student-card-skeleton";
+
+// Page-shell skeleton for the OGS-groups gate/Suspense states.
+export function OgsGroupsPageSkeleton() {
+  return (
+    <StudentCardPageSkeleton
+      label="Gruppe wird geladen"
+      testId="ogs-groups-skeleton"
+    />
+  );
+}
