+++ frontend/src/app/[tenant]/(protected)/staff/loading.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { StaffPageSkeleton } from "./page-skeleton";
+
+/**
+ * Route-level loading UI: renders the same skeleton the page shows while its
+ * data loads, so navigation shows one continuous skeleton instead of the
+ * generic group-level Loading followed by the page skeleton.
+ */
+export default function StaffLoading() {
+  return <StaffPageSkeleton />;
+}
